requirements.txt
pygame>=2.5.2
numpy>=1.24.3
google-generativeai>=0.3.0
